Conflicts:
	code/playground/src/aco/optimizer2.cpp

+++ scripts/llvm_install.sh
@@ -1,22 +1,36 @@
 # setup environment variables
-. environment.setup
+. ./environment.setup
 
 VERSION=3.0
+
+rm -R $PREFIX/llvm-$VERSION
 
 # download llvm 
 echo "*****************************************"
 echo "* Downloading current LLVM distribution *"
 echo "*****************************************"
-wget http://llvm.org/releases/$VERSION/llvm-$VERSION.tar.gz 
+wget -nc http://llvm.org/releases/$VERSION/llvm-$VERSION.tar.gz 
+
+RET=$?
+if [ $RET -ne 0 ]; then
+	exit $RET
+fi
+
 tar -xf llvm-$VERSION.tar.gz
-# change into tools
+# change dire into tools
 cd llvm-$VERSION.src/tools
 
 echo "******************************************"
 echo "* Downloading current CLANG distribution *"
 echo "******************************************"
 # download clang
-wget http://llvm.org/releases/$VERSION/clang-$VERSION.tar.gz 
+wget -nc http://llvm.org/releases/$VERSION/clang-$VERSION.tar.gz 
+
+RET=$?
+if [ $RET -ne 0 ]; then
+	exit $RET
+fi
+
 tar -xf clang-$VERSION.tar.gz
 mv clang-$VERSION.src clang
 rm -f clang-$VERSION.tar.gz
@@ -25,34 +39,43 @@ cd ../
 echo "***********************************"
 echo "* Applying insieme patch to CLANG *"
 echo "***********************************"
-patch -p1  < ../insieme-$VERSION.patch
+patch -p1  < ../patches/insieme-$VERSION.patch
 
 echo "*******************"
 echo "* Compiling CLANG *"
 echo "*******************"
-./configure --prefix=$PREFIX/llvm-$VERSION --enable-shared=yes --enable-assert=yes --enable-debug-runtime=no \
---enable-debug-symbols=no --enable-optimized=yes
+
+export LD_LIBRARY_PATH=$PREFIX/gcc-latest/lib64:$PREFIX/gmp-latest/lib:$PREFIX/mpfr-latest/lib:$PREFIX/cloog-gcc-latest/lib:$PREFIX/ppl-latest/lib:$PREFIX/mpc-latest/lib/:$LD_LIBRARY_PATH 
+
+CFLAGS="-mtune=native -O3 -fgraphite-identity"
+CC=$CC CXX=$CXX CFLAGS=$CFLAGS CXXFLAGS=$CFLAGS LDFLAGS="-mtune=native -O3" ./configure --prefix=$PREFIX/llvm-$VERSION --enable-shared=yes \
+  	 --enable-assert=yes --enable-debug-runtime=no --enable-debug-symbols=no --enable-optimized=yes
 # --enable-doxygen=yes
+
 make REQUIRES_RTTI=1 clang-only -j$SLOTS
 
+# Check for failure
+RET=$?
+if [ $RET -ne 0 ]; then
+	exit $RET
+fi
 
-echo "********************"
-echo "* Installing CLANG *"
-echo "********************"
 make clang-only install
+
+cd ../
+echo "****************************************"
+echo "* Removing LLVM installation directory *"
+echo "****************************************"
+rm -R llvm-$VERSION.src
+rm -f llvm-$VERSION.tar.gz
+
 
 #echo "****************************************************************"
 #echo "* Patching stdarg.h to make CLANG work with linux libc (maybe) *"
 #echo "****************************************************************"
-cd ../
-patch -d $PREFIX/llvm-$VERSION/lib/clang/$VERSION/include < stdarg.patch
+patch -d $PREFIX/llvm-$VERSION/lib/clang/$VERSION/include < ./patches/stdarg.patch
 
-rm $PREFIX/llvm-latest
+rm -f $PREFIX/llvm-latest
 ln -s $PREFIX/llvm-$VERSION $PREFIX/llvm-latest
 
-echo "****************************************"
-echo "* Removing LLVM installation directory *"
-echo "****************************************"
-rm -rf llvm-$VERSION.src
-rm -rf llvm-$VERSION.tar.gz
-
+exit 0
